==> notebooks/matching_pipeline/01_clean_eu_sanctions.ipynb <==
The goal of this notebook is to clean and standardize the raw EU Financial Sanctions dataset, preparing it for future name‑matching experiments.

The notebook is structured as follows:

- **Preprocessing**: Selecting relevant columns and filtering the dataset

- **Language Normalization**: Handling inconsistencies across Latin‑ and non‑Latin‑based names

- **Name Cleaning**: Formatting names for consistency 

- **Outlier Analysis** Removing the top 1% of entities with too many names to reduce noise

- **Data Aggregation** Creating a condensed version of the cleaned dataset, with a single row per entity


# 1. Imports

#necessary libraries
from pathlib import Path
import pandas as pd  
import numpy as np  
import warnings  
from unidecode import unidecode
import re  
import matplotlib.pyplot as plt
import seaborn as sns


#commands for better ouput readability 
pd.set_option('display.max_columns', None)   
#pd.set_option('display.max_rows', None)   
warnings.filterwarnings("ignore", category=UserWarning, module='pandas')  

# 2. Configuration

#paths
project_dir=Path.cwd().parent.parent
raw_dir=project_dir/'data'/'raw'
processed_dir=project_dir/'data'/'processed'
processed_dir.mkdir(exist_ok=True)  #keep this just in case :/

eu_file=raw_dir/'eu_condensed_List.csv'

df=pd.read_csv(eu_file, delimiter=';')

# 3. Preprocessing

df.head()

The raw dataset contains extensive information about each entity, including names, addresses, birthplaces, citizenship, and other details. Since this project focuses specifically on exploring name matching, the scope was narrowed down to the name fields and relevant sanction-related information.

#extract the relevant columns from the original dataframe into df_names
df_names=df[['Entity_LogicalId','Entity_EU_ReferenceNumber','Entity_SubjectType_ClassificationCode','Entity_Regulation_EntryIntoForceDate','Entity_Regulation_NumberTitle','Entity_Regulation_Programme','NameAlias_WholeName','NameAlias_NameLanguage']]

#renamecolumns into more concise and intuitive names 
df_names.columns=['Id','EU Reference Number','Entity Type','Entry Into Force Date','Regulation Identifier','Sanction Programme','Name','Name Language']

#delete rows that have become redundant after deleting columns 
df_names=df_names.replace('', np.nan)
df_names=df_names.dropna(subset=['Name'])  

df_names=df_names.reset_index(drop=True)

df_names.head()

# 4. Language Normalization

df_names['Name Language'].unique()

By reviewing the names listed for each language, it was possible to categorize them into two groups: Latin-based and Non-Latin-based languages.

- Latin-based languages: These languages use the Latin alphabet, but may include diacritical marks.
   - **Danish (DA):** æ,å,ø
   - **Polish (PO):** ł,ż,ę
   - **French (FR):** ï,é,œ
- Non-Latin-based languages: These languages have their own unique alphabets.
   - **Arabic (AR):** عبد المنان آغا
   - **Russian (RU):** Виктор Владимирович
   - **Hebrew (HE):** מאיר אטינגר

### 4.1 **Non-latin Languages**

For simplicity, all rows with names in non‑Latin scripts were removed from the dataset. This is because these names cannot be reliably transliterated into English with a simple rule‑based approach, and matching them accurately would require advanced linguistic expertise. The excluded languages are:
- Greek (EL)
- Bulgarian (BG) 
- Russian (RU)
- Farsi (FA)
- Korean (KO)
- Ukranian (UK)
- Chinese (ZH)
- Pashto (PS)
- Serbian (SR)
- Hebrew (HE)

#correcting a data error: 'Bahram Hosseini MOTLAGH' was mislabelled as Farsi
df_names.loc[df_names['Name']=='Bahram Hosseini MOTLAGH','Name Language']=np.nan

#list of non-latin languages
non_latin_list=['EL','BG','RU','FA','KO','UK','ZH','PS','SR','HE']  

#filter out rows corresponding to non-latin languages 
df_names=df_names[~df_names['Name Language'].isin(non_latin_list)]
df_names=df_names.reset_index(drop=True)    


### 4.2 **Latin-based Languages**

For the remaining languages, diacritics and special characters were handled using `unidecode`. This library intelligently maps non-ASCII Latin characters to their closest ASCII equivalents (e.g., ø → o, æ → ae, ç → c). More details about its limitations can be found at [Unidecode on PyPI](https://pypi.org/project/Unidecode/).


def normalize(text):
    """
    Normalizes text by removing diacritics and special characters.

    Args:
        text (str): The input text to be normalized.

    Returns:
        str: The normalized text with diacritics and special characters removed.

    Example:
        >>> normalize('ołá')
        'ola'
    """
    return unidecode(text)

df_names['Name']=df_names['Name'].apply(normalize)

#drop the 'Name Language' column (no longer needed)
df_names = df_names.drop(columns=['Name Language'])

# 5. Name Cleaning

After inspecting the names in the dataset, we noticed recurring inconsistencies in their formatting. To standardize them, we created two cleaning functions: `clean_name_person`and `clean_name_enterprise`. Both functions apply the same text normalization and cleaning rules, differing only in their handling of numerical characters (relevant for enterprises). 

def clean_name_person(text):
    """
    Cleans an individual's name using regex patterns.

    Rules:
    - Replaces apostrophes and hyphens with spaces
    - Removes any character that is not a letter or space
    - Capitalizes the first letter of each word

    Args:
        text (str): The raw name.

    Returns:
        str: The cleaned, properly capitalized name.
    """
    
    text=re.sub(r"[\'\-\\]",' ',text)  
    text=re.sub(r"[^a-zA-Z\s]",'',text) 
    
    text=text.split()
    text=[word.capitalize() for word in text]  

    text=' '.join(text)
    
    return text

#apply the cleaning function to 'person' rows
df_names.loc[df_names['Entity Type']=='person','Name']=df_names.loc[df_names['Entity Type']=='person','Name'].apply(clean_name_person)

def clean_name_enterprise(text):
    """
    Cleans an enterprise name using regex patterns.

    Rules:
    - Replaces apostrophes and hyphens with spaces
    - Removes any character that is not a letter, digit, or space
    - Capitalizes the first letter of each word

    Args:
        text (str): The raw enterprise name.

    Returns:
        str: The cleaned, properly capitalized name.
    """
    
    text=re.sub(r"[\'\-\\]",' ',text)  
    text=re.sub(r"[^a-zA-Z0-9\s]",'',text)
    
    text=text.split()
    text=[word.capitalize() for word in text]  

    text=' '.join(text)
    
    return text

#apply the cleaning function to 'enterprise' rows
df_names.loc[df_names['Entity Type']=='enterprise','Name']=df_names.loc[df_names['Entity Type']=='enterprise','Name'].apply(clean_name_enterprise)

#run this to avoid errors
df_names=df_names.drop_duplicates().reset_index(drop=True)

df_names.head()

# 6. Outlier Analysis

Names associated with the same entity can vary significantly, from a handful of entries to 100+. To prevent disproportionately long or ambiguous records from affecting the quality of name matching, we identified and removed outliers:

- We first group names by entity (Id) and count how many names each entity had.

- We observed a highly skewed distribution, with certain entities having an unusually high number of names.

- To formalize the cutoff, we computed the 99th percentile of counts for people and enterprises:

- All entities with counts above these thresholds were treated as outliers and removed from the dataset.





#print Name Counts per ID

person_counts=df_names[df_names['Entity Type']=='person'].groupby('Id').size().reset_index(name='Count')
enterprise_counts=df_names[df_names['Entity Type']=='enterprise'].groupby('Id').size().reset_index(name='Count')

person_counts=person_counts.sort_values(by='Count',ascending=False)
enterprise_counts=enterprise_counts.sort_values(by='Count',ascending=False)

print(person_counts)
print(enterprise_counts)

#print 99th percentile name count

x_p=person_counts['Count']
x_e=enterprise_counts['Count']

upper_p_new=np.percentile(x_p, 99) 
upper_e_new=np.round(np.percentile(x_e, 99), 1) 

print('Upper limit for People:',upper_p_new)
print('Upper limit for Enterprises:',upper_e_new)

#remove outliers from dataset

outliers_person=person_counts[person_counts['Count']>upper_p_new].reset_index(drop=True)
outliers_enterprise=enterprise_counts[enterprise_counts['Count']>upper_e_new].reset_index(drop=True)

outliers_Id=pd.concat([outliers_person['Id'], outliers_enterprise['Id']], ignore_index=True).drop_duplicates()
df_names=df_names[~df_names['Id'].isin(outliers_Id)].reset_index(drop=True)

# 7. Data Agreggation

For future experiments, we created a condensed version of the dataset, `df_grouped`, by aggregating all names associated with a single ID into a single cell. This approach will potentially allow for more efficient processing and matching.

df_grouped=df_names.copy()

def aggregate_unique_words(grouped_names):
    """

    This function takes a list of names, splits each name into its individual words,
    and returns a comma-separated string of all unique words across the list.

    Args:
        grouped_names (list of str): List of names associated with a single entity.

    Returns:
        str: A comma-separated string of unique words extracted from the names.
    """
 
    words=set()

    for name in grouped_names:

        #example: Abu Ali - [Abu,Ali]
        name=name.split()
        words.update(name)
        
    return ', '.join(words)

            
df_grouped=df_grouped.groupby('Id',as_index=False).agg({
       
    'EU Reference Number': 'first',  
    'Entity Type': 'first',
    'Entry Into Force Date': 'first',
    'Regulation Identifier': 'first',
    'Sanction Programme': 'first',
    'Name': aggregate_unique_words,  
})

# 8. Output

df_names.info()


df_grouped.info()

df_names.to_pickle(processed_dir/'cleaned_eu_sanctions.pkl')
df_grouped.to_pickle(processed_dir/'cleaned_eu_sanctions_grouped.pkl')
