==> notebooks/matching_pipeline/03_match_names_standard.ipynb <==
The goal of this notebook is to match UK sanctioned individuals to EU sanctioned individuals using a standard fuzzy name-matching approach.
This approach serves as a baseline method for comparison with the more advanced grouped and token‑based technique developed later.

The notebook is structured as follows:  

- **Preprocessing**: Standardizing dataframes for consistent format
  
- **Fuzzy Matching**: Uisng the `rapidfuzz`library to find the best EU match for every UK name. 

- **Labelling**:  Assigning final labels to matched pairs (`match`, `not match`, `preliminary match`)   


# 1. Imports

#necessary libraries
from pathlib import Path
import pandas as pd  
import numpy as np  
import warnings  
from unidecode import unidecode
import re  
import matplotlib.pyplot as plt
import seaborn as sns
import time
from rapidfuzz import process, fuzz
import random
import scipy.stats as st
import math
 
#commands for better output readability 
pd.set_option('display.max_columns', None)  #display all columns 
pd.set_option('display.max_rows', None)  #display all rows 
warnings.filterwarnings("ignore", category=UserWarning, module='pandas')  #surpress warning messages

# 2. Configuration

#paths
project_dir=Path.cwd().parent.parent
processed_dir=project_dir/'data'/'processed'
final_dir=project_dir/'data'/'final'

eu_processed_file=processed_dir/'cleaned_eu_sanctions.pkl'
uk_processed_file=processed_dir/'cleaned_uk_sanctions.pkl'

df_eu=pd.read_pickle(eu_processed_file)  
df_uk=pd.read_pickle(uk_processed_file)  

# 3. Preprocessing

df_eu.head()

df_uk.head()

#harmonize EU/UK column names for consistency 
df_eu=df_eu.rename(columns={'Name':'EU Name','Sanction Programme':'EU Sanction Programme','Regulation Identifier':'EU Regulation Identifier','Id':'EU ID'})
df_uk=df_uk.rename(columns={'Name':'UK Name','Group ID': 'UK ID','Regime':'UK Sanction Programme'})

#keep only entries of individual people
df_eu=df_eu[df_eu['Entity Type']!='enterprise']
df_uk=df_uk[df_uk['Group Type']!='Entity']

df_eu=df_eu.drop(columns=['Entity Type','Entry Into Force Date']).reset_index(drop=True)
df_uk=df_uk.drop(columns=['Group Type','Last Updated']).reset_index(drop=True)

df_eu.head()


df_uk.head()

# 4. Fuzzy Matching

Fuzzy matching measures how closely two strings resemble each other, allowing for inexact or approximate matches. This is especially useful when dealing with names that may have variations in spelling, ordering, or transcription. In this project, we used the `rapidfuzz` library, and its `fuzz` module, to assess similarity between UK and EU names.

We experimented with all available scorers on a sample of UK names to assess match quality. Results showed:

`fuzz.ratio` and `fuzz.partial_ratio` performed well for direct matches but were prone to error when token order varied.

`fuzz.partial_token_set_ratio` caused significant false positives despite higher scores.

`fuzz.token_sort_ratio` offered the best balance and fewer false positives.

Decision: We selected `fuzz.token_sort_ratio` as the primary metric for the matching process.

def match_names(name):
    """
    Finds the best match for a UK name from the EU names list using fuzz.token_sort_ratio.

    Args:
        name (str): The UK name to match against the EU names.

    Returns:
        pd.Series: A Series containing:
            - match (str): The best EU name match.
            - score (float): The rounded match score (0–100).
            - eu_id (str): The associated EU ID for the best match.
    """

    match,score,index=process.extractOne(name, df_eu['EU Name'].tolist(),scorer=fuzz.token_sort_ratio)
    eu_id=df_eu.iloc[index]['EU ID'] 
    return pd.Series([match,int(round(score)),eu_id])


df_uk[['EU Name Match','Score','EU ID']]=df_uk['UK Name'].apply(match_names)


df_uk.head(10)

### Final Candidate Selection
At this point, the UK dataset contains one or more candidate EU matches per UK ID. To reduce these to a single best match per UK ID, we apply the following hierarchical criteria:

1. Highest Score: Select the candidate with the highest match score

2. Longest Name: If there are ties in the score, select the candidate with the longest UK Name (longer names are more likely to reliably identify an individual)

3. First Appearance:  In the rare case of ties in both score and length, keep the candidate that appears first

df_uk=df_uk.sort_values(by='UK ID',ascending=True).reset_index(drop=True)

#keep rows with highest scores per UK ID
df_uk['Max Score'] = df_uk.groupby('UK ID')['Score'].transform('max')
df_uk=df_uk[df_uk['Score']==df_uk['Max Score']].reset_index(drop=True)

#keep rows with longest names per UK ID
df_uk['String Length']=df_uk['UK Name'].str.len()
df_uk['Max String Length']=df_uk.groupby('UK ID')['String Length'].transform('max')
df_uk=df_uk[df_uk['Max String Length']==df_uk['String Length']].reset_index(drop=True)

df_uk = df_uk.drop_duplicates(subset='UK ID', keep='first').reset_index(drop=True)

df_uk=df_uk.drop(columns=['Max String Length','Max Score'])

df_uk.head()

#filtering matches with a specific score to analyse trends in accuracy

df_uk.loc[df_uk['Score']==70,['UK ID','UK Name','EU Name Match']]  #iteratively adjust score

# 5. Labelling Results

Financial institutions must reliably flag sanctioned individuals to comply with Financial Sanctions. Due to the sensitive nature of this task, human intervention is necessary to ensure accuracy, as the stakes are high when mistakenly targeting innocent individuals. To balance these needs, our strategy focused on minimizing manual intervention by labeling matched pairs as `Match`, `Not Match`, or `Preliminary Match`. Confident outcomes, either strong matches or clear non-matches, were assigned to Match and Not Match, respectively. Ambiguous cases fell under Preliminary Match, a buffer category reserved for human review. The labeling thresholds were determined empirically by inspecting the distribution of match scores and identifying natural cutoffs.

def add_label(score):

    if score<70:
        label='not match'
    elif score>=91:
        label='match'
    else:
        label='preliminary match'
    return label

df_uk['Label']=df_uk['Score'].apply(add_label)

# 6. Output

### 6.1 For further analysis

df_uk.head()

df_uk.info()

#just some final tweaks for presentation and consistency
column_order=['UK ID','UK Sanction Programme','UK Name','EU Name Match','EU ID','Score','String Length','Label']
df_uk=df_uk[column_order]

df_uk.to_pickle(processed_dir/'matched_standard.pkl')

### 6.2 For final output

df_output=df_uk.copy()
df_output=df_output.drop(columns=['String Length'])

df_output.head()

df_output.to_csv(final_dir/'matched_names_standard_method.csv',index=False)
